rename Led5050 component file
