+++ lib/index.ts
@@ -72,4 +72,4 @@ export * from "./MountedPcbModule"
 export * from "./Screen"
 export * from "./JSTZH1_5mm"
 export * from "./FPC"
-export * from "./LED5050"
+export * from "./Led5050"
